Implemented Multiple UI Buttons for AR Player
Implemented Multiple UI Buttons for AR Player.

Object Selection and placement buttons:
Have now successfully tied the creation of chest and trap objects to the UI for the AR Player, first the button is selected, a section and then confirm button to spawn in an object of this type onto the map, may look into refactoring this into a single function for efficiency.

Map Scaling Slider:
Slider added to the AR Players start screen in order to scale the map before the game starts.

Map Rotation Buttons:
A left and right rotation button have been added these buttons will allow the player to rotate the map to face in front of them before the game starts, may keep them as could be useful.