Entire Revamp of the Turn System and Fixing issues with the Buttons
Th turn system is now entirely controlled by network messages, we'll see if that's a good idea later. The issue with button presses leading to crashing seems to be fixed by changing regular pointers on the Observer array in Subject to a TWeakPointer, again we'll see whether this has entirely stopped the problem, from the looks of things they're getting tossed in during garbage collection, very odd.