Refactor of Entity Pool
So due to the need of separation between static and "living" entities, one having to derive from Actor and the other from Character, the decision was therefore made to create PooledEntityComponent. The component has been created using an interface, this component will be attached to entities that should exist within the pools, this has all been confirmed to work.