Setup Widgets for the AR and Start
Setup Widgets for AR Player, all that will expand when they're clicked to reveal an explanation of their function, for better communication and some bug fixing on the map section selection.

A starting area has been added outside of the dungeon for the VR Player to wait in before the AR Player has finished their setup.

Starting setup for the AR Player has been implemented allowing them to place down 5 traps and 5 treasures before the game starts with the number of objects they have left being displayed to them.